feat: port in more from the C++ code (#24)
This PR ports in some more functionality based on the MIT-licensed C++
code from CWI.

In particular, it implements the following:

* The `makeSample` function from C++ to build a sample of ~16KB from the
input data
* The `suffix limit` optimization and its corresponding `finalize`
method needed when building the symbol table, including changes to the
`compress_word` function we have that more directly corresponds to the
`compressVariant` from the C++ code
* The `byteCodes` from C++, which we implement here as `codes_one_byte`.
Note that before this PR, one-byte codes would not be found unless the
byte occurred at the end of the plaintext string
* Separates the `Compressor` build state into a new `CompressorBuilder`
struct, which has all methods that take `&mut self`. This also means
that we can in theory construct a `Compressor` now from a symbol table,
though that logic is not implemented.

Additional things in this PR:

* Added a micro benchmark for `compress_word` method comparing the
relative speeds of both code paths, see
https://github.com/spiraldb/fsst/pull/24#discussion_r1742263401
* Removed many of the old small-data benchmarks. I've added several of
the `dbtext` compression benchmarks from the CWI paper. Here's a table
of the compression factors:

dbtext | c++ compress factor | fsst-rs compress factor
-------|-----|-------
l_comment | 2.73 | 2.69
urls | 2.33 | 2.27
wikipedia | 1.81 | 1.75

I'll follow up to figure out how to close the gap with those 1-2%
differences